pyproject.toml
[project]
name = "flask-boilerplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic==1.15.1",
    "blinker==1.7.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "flask==3.0.2",
    "flask-marshmallow==1.2.0",
    "flask-migrate==4.0.5",
    "flask-seasurf>=2.0.0",
    "flask-sqlalchemy==3.1.1",
    "flask-talisman>=1.1.0",
    "greenlet==3.1.1",
    "gunicorn==21.2.0",
    "iniconfig==2.0.0",
    "itsdangerous==2.1.2",
    "jinja2==3.1.3",
    "mako==1.3.2",
    "markupsafe==2.1.5",
    "marshmallow==3.26.1",
    "marshmallow-sqlalchemy==1.4.1",
    "mccabe==0.7.0",
    "packaging==23.2",
    "pluggy==1.4.0",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.38",
    "typing-extensions==4.12.2",
    "werkzeug==3.0.1",
]

[dependency-groups]
dev = [
    "coverage==7.4.3",
    "flake8==7.0.0",
    "pycodestyle==2.11.1",
    "pyflakes==3.2.0",
    "pytest==8.0.0",
]
